+++ src/Pages/Accounts/Components/AccountButtons/Components/index.js
@@ -2,3 +2,4 @@ export { default as AccountAssets } from './AccountAssets';
 export { default as AccountDelegations } from './AccountDelegations';
 export { default as AccountRewards } from './AccountRewards';
 export { default as AccountUnbondings } from './AccountUnbondings';
+export { default as AccountAttributes } from './AccountAttributes';