+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add account delegations/unbondings to account page #218
 - Add pagination to account assets #224
 - Add vote indicator for block validators #167
+- Add account attributes #244
 
 ## 2.5.0